fixed some functions
