+++ contracts/FixedMath.sol
@@ -1,4 +1,4 @@
-pragma solidity ^0.7;
+pragma solidity ^0.6.8;
 
 // Some inspiration were taken from FixidityLib, but stripped down as overflow is most likely not going to happen
