+++ contracts/FixidityLib.sol
@@ -1,4 +1,4 @@
-pragma solidity ^0.5.0;
+pragma solidity ^0.6.8;
 
 
 /**
